// src/typings.d.ts
export interface InputGetValueMethod extends BaseMethod {
  method: 'getValue';
  success?: Callback<{
    /**
     * Input content
     * @Android
     * @iOS
     * @Harmony
     * @Web
     * @since 3.4
     */
    value: string;
    /**
     * Begin position of the selection
     * @Android
     * @iOS
     * @Harmony
     * @Web
     * @since 3.4
     */
    selectionStart: number;
    /**
     * End position of the selection
     * @Android
     * @iOS
     * @Harmony
     * @Web
     * @since 3.4
     */
    selectionEnd: number;
    /**
     * Is composing or not, iOS only
     * @Android
     * @iOS
     * @Harmony
     * @Web
     * @since 3.4
     */
    isComposing: boolean;
  }>;
}

/**
 *  Set input content
 * @Android
 * @iOS
 * @Harmony
 * @Web
 * @since 3.4
 */
export interface InputSetValueMethod extends BaseMethod {
  method: 'setValue';
  params: {
    /**
     * Input content
     * @Android
     * @iOS
     * @Harmony
     * @Web
     * @since 3.4
     */
    value: string;
  };
}
// ...
export interface TextAreaGetValueMethod extends BaseMethod {
  method: 'getValue';
  success?: Callback<{
    /**
     * Input content
     * @Android
     * @iOS
     * @Harmony
     * @Web
     * @since 3.4
     */
    value: string;
    /**
     * Begin position of the cursor
     * @Android
     * @iOS
     * @Harmony
     * @Web
     * @since 3.4
     */
    selectionStart: number;
    /**
     * End position of the cursor
     * @Android
     * @iOS
     * @Harmony
     * @Web
     * @since 3.4
     */
    selectionEnd: number;
    /**
     * Is composing or not, iOS only
     * @iOS
     * @Android
     * @Harmony
     * @Web
     * @since 3.4
     */
    isComposing: boolean;
  }>;
}

/**
 * Set input content
 * @Android
 * @iOS
 * @Harmony
 * @Web
 * @since 3.4
 */
export interface TextAreaSetValueMethod extends BaseMethod {
  method: 'setValue';
  params: {
    /**
     * Input content
     * @Android
     * @iOS
     * @Harmony
     * @Web
     * @since 3.4
     */
    value: string;
  };
}